+++ SieveOfEratosthenes/changelog.md
@@ -2,3 +2,4 @@
 * fix time calculation
 * skip even numbers
 * allocate result on stack
+* use std::vector<bool> instead of std::bitset